requirements.txt
requests>=2.31,<3
python-dotenv>=1.0,<2
mysql-connector-python>=9.0,<10
pyyaml>=6.0,<7
